Cargo.toml: mark project as MIT
The crates.io website has trouble finding out the license of the
project, because the LICENSE.md file contains your name inside of it and
thus is non-usual.

By directly specifying that the project is under the MIT license in the
Cargo.toml file, we can help crates.io display the proper license on the
crate's page.