add support for "*.ui" files, use regex
In the official GTK documentation, the .ui file extension is used to
contain the XML payload describing the user interface. Meanwhile, glade
is saving its files using the .glade file extension.

Both extensions should be supported.

A new dependency to the `regex` crate was introduced, as it was
suggested by a comment in the code. This allows to match and replace
both file extensions.

A dependency to the `lazy_static` crate has been introduced as well, in
order to compile the regex only once during startup, and not every time
the generate_bind_recursive() function is called.

A separate function has been created in order to be able to write some
unit tests to verify its behavior. It should be inlined by the compiler,
leading to zero additional cost.